Export handler and use adjustHeight
